backend/requirements.txt
# Common Python requirements for the research project backend
# Add specific requirements for each component in their subfolders

Flask==2.3.2
werkzeug==2.3.6
requests==2.31.0

# API server
fastapi>=0.110
uvicorn[standard]>=0.24
python-multipart>=0.0.6
aiofiles>=23.2.1

# Data ingestion
pandas>=2.1
openpyxl>=3.1
PyPDF2>=3.0
python-docx>=1.1

# NLP / ML
spacy>=3.7
scikit-learn>=1.3
joblib>=1.3

# Image Processing & Computer Vision
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
numpy>=1.24.4
pillow>=10.2.0

# Machine Learning
torch
PyYAML>=6.0